Testing/ilpv4 enhancements (#167)
* Updates for routing…
* Fix Codecs for CCP
* Fixes #162
* Only strip leading bytes if the byte-length is < 1
* Add unit test for AsnUintCodec
* Fix encoding of short vs integer
* encode/decode `short` values with AsnUint8Codec
* encode/decode `int` values with AsnUint16Codec
* Fix unit tests.
* Improve Javadoc and formatting.
* Remove duplicate code.

+++ codec-framework/src/test/java/org/interledger/encoding/asn/framework/OnValueChangedTest.java
@@ -9,9 +9,9 @@ package org.interledger.encoding.asn.framework;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,7 +20,6 @@ package org.interledger.encoding.asn.framework;
  * =========================LICENSE_END==================================
  */
 
-import org.interledger.encoding.asn.codecs.AsnSequenceCodec;
 import org.interledger.encoding.asn.codecs.AsnUint8Codec;
 
 import org.hamcrest.MatcherAssert;
@@ -32,14 +31,14 @@ public class OnValueChangedTest {
   @Test
   public void testEvent() {
 
-    final int[] values = new int[]{0};
+    final int[] values = new int[] {0};
 
     AsnUint8Codec codec = new AsnUint8Codec();
     codec.setValueChangedEventListener(source -> {
       values[0] = source.decode();
     });
 
-    codec.encode(1);
+    codec.encode((short) 1);
 
     MatcherAssert.assertThat("Event was fired.", values[0] == 1);
 
@@ -49,8 +48,10 @@ public class OnValueChangedTest {
   public void testSetterWithExistingListener() {
 
     AsnUint8Codec codec = new AsnUint8Codec();
-    codec.setValueChangedEventListener(source -> { });
-    codec.setValueChangedEventListener(source -> { });
+    codec.setValueChangedEventListener(source -> {
+    });
+    codec.setValueChangedEventListener(source -> {
+    });
 
   }
 
@@ -58,7 +59,8 @@ public class OnValueChangedTest {
   public void testRemove() {
 
     AsnUint8Codec codec = new AsnUint8Codec();
-    codec.setValueChangedEventListener(source -> { });
+    codec.setValueChangedEventListener(source -> {
+    });
     codec.removeEncodeEventListener();
 
     MatcherAssert.assertThat("Listener was removed.", !codec.hasValueChangedEventListener());
